chore: refactor gateway Form components
chore: fix
