+++ apps/router/src/gateway-ui/components/walletCard/LightningCard.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import { Box, Button, Flex, Text } from '@chakra-ui/react';
 import { useTranslation } from '@fedimint/utils';
 import { FaArrowDown, FaArrowUp } from 'react-icons/fa';
@@ -13,21 +13,24 @@ export const LightningCard = React.memo(function LightningCard(): JSX.Element {
   const { t } = useTranslation();
   const { state, dispatch } = useGatewayContext();
 
-  const handleModalOpen = (action: WalletModalAction) => {
-    const federations = state.gatewayInfo?.federations;
-    if (!federations?.length) return;
+  const handleModalOpen = useCallback(
+    (action: WalletModalAction) => {
+      const federations = state.gatewayInfo?.federations;
+      if (!federations?.length) return;
 
-    dispatch({
-      type: GATEWAY_APP_ACTION_TYPE.SET_WALLET_MODAL_STATE,
-      payload: {
-        action,
-        type: WalletModalType.Lightning,
-        selectedFederation: federations[0],
-        showSelector: true,
-        isOpen: true,
-      },
-    });
-  };
+      dispatch({
+        type: GATEWAY_APP_ACTION_TYPE.SET_WALLET_MODAL_STATE,
+        payload: {
+          action,
+          type: WalletModalType.Lightning,
+          selectedFederation: federations[0],
+          showSelector: true,
+          isOpen: true,
+        },
+      });
+    },
+    [dispatch, state.gatewayInfo?.federations]
+  );
 
   return (
     <Box p={6} bg='white' borderRadius='lg' borderWidth='1px' mb={4}>
