+++ apps/router/src/languages/en.json
@@ -483,7 +483,10 @@
     },
     "send": {
       "submit": "Submit",
+      "lightning-invoice": "Lightning Invoice",
+      "lightning-invoice-placeholder": "lnbc...",
       "to-onchain-address": "To Onchain Address",
+      "pay-invoice": "Pay Lightning Invoice",
       "address-placeholder": "bc1p...",
       "peg-out-to-onchain": "Peg Out Ecash to Onchain",
       "peg-out-success": "Peg Out TX Sent!",
